bitvmx_client::errors
Enum BitVMXError

pub enum BitVMXError {

    DisputeResolutionProtocolSetup(String),
    ConfigurationError(ConfigError),
    InvalidParameter(String),
    ParseIntError(ParseIntError),
    KeyManagerError(KeyManagerError),
    BitcoinError(BitcoinClientError),
    UnspendableKeyError(UnspendableKeyError),
    ProtocolBuilderError(ProtocolBuilderError),
    FromHexError(FromHexError),
    StorageError(StorageError),
    ProgramNotFound(Uuid),
    ProtocolNotFound(String),
    ProgramError(ProgramError),
    WinternitzError(WinternitzError),
    ProgramNotReady(Uuid),
    VariableNotFound(Uuid, String),
    InvalidVariableType(String),
    InvalidWitnessType,
    DispatcherError(DispatcherError),
    CommsEncodingError(OperatorCommsError),
    CommsCommunicationError,
    InvalidCommsAddress(String),
    KeysNotFound(Uuid),
    BitcoinCoordinatorError(BitcoinCoordinatorError),
    BrokerError(BrokerError),
    SerdeSerializationError(Error),
    InvalidMsgVersion,
    InvalidMessageType,
    SerializationError,
    InvalidMessageFormat,
    InvalidTransactionName(String),
    MessageProcessingError,
    MuSig2SignerError(Musig2SignerError),
    ProgramAlreadyExists(Uuid),
    ScriptError(ScriptError),
    ClientError(ClientError),
    NotImplemented(String),
    EmulatorError(EmulatorError),
    EmulatorResultError(EmulatorResultError),
    ProgramDefinitionError(ProgramDefinitionError),
    InstructionNotFound(String),
    ChallengeNotFound(String),
    ChallengeIdxNotFound(u32),
    RomDataNotFound(u32),
    InsufficientAmount,
    TransactionNotFoundInBlock,
    InconsistentZKPData(Uuid),
    DirectoryCreationError(String, Error),
    InvalidParticipant(String),
    MissingPartialSignatures(String, String),
    MissingPublicNonces(String, String),
    WalletError(WalletError),
    InvalidList(String),
    IdentificationError(IdentificationError),
    JobDispatcherNotResponding(String),

}

Variants

DisputeResolutionProtocolSetup(String)

ConfigurationError(ConfigError)

InvalidParameter(String)

ParseIntError(ParseIntError)

KeyManagerError(KeyManagerError)

BitcoinError(BitcoinClientError)

UnspendableKeyError(UnspendableKeyError)

ProtocolBuilderError(ProtocolBuilderError)

FromHexError(FromHexError)

StorageError(StorageError)

ProgramNotFound(Uuid)

ProtocolNotFound(String)

ProgramError(ProgramError)

WinternitzError(WinternitzError)

ProgramNotReady(Uuid)

VariableNotFound(Uuid, String)

InvalidVariableType(String)

InvalidWitnessType

DispatcherError(DispatcherError)

CommsEncodingError(OperatorCommsError)

CommsCommunicationError

InvalidCommsAddress(String)

KeysNotFound(Uuid)

BitcoinCoordinatorError(BitcoinCoordinatorError)

BrokerError(BrokerError)

SerdeSerializationError(Error)

InvalidMsgVersion

InvalidMessageType

SerializationError

InvalidMessageFormat

InvalidTransactionName(String)

MessageProcessingError

MuSig2SignerError(Musig2SignerError)

ProgramAlreadyExists(Uuid)

ScriptError(ScriptError)

ClientError(ClientError)

NotImplemented(String)

EmulatorError(EmulatorError)

EmulatorResultError(EmulatorResultError)

ProgramDefinitionError(ProgramDefinitionError)

InstructionNotFound(String)

ChallengeNotFound(String)

ChallengeIdxNotFound(u32)

RomDataNotFound(u32)

InsufficientAmount

TransactionNotFoundInBlock

InconsistentZKPData(Uuid)

DirectoryCreationError(String, Error)

InvalidParticipant(String)

MissingPartialSignatures(String, String)

MissingPublicNonces(String, String)

WalletError(WalletError)

InvalidList(String)

IdentificationError(IdentificationError)

JobDispatcherNotResponding(String)

Implementations

impl BitVMXError

pub fn is_fatal(&self) -> bool

Returns whether the error should be considered fatal.

NOTE: For now, we mark all errors as fatal to avoid masking issues and ensure coordinated shutdown paths are exercised. Once the system is stable, we should revisit this and separate truly fatal errors (e.g., storage corruption) from recoverable ones.