bitvmx_client::program::program
Struct Program

pub struct Program {
    pub program_id: Uuid,
    pub my_idx: usize,
    pub participants: Vec<ParticipantData>,
    pub leader: usize,
    pub protocol: ProtocolType,
    pub state: ProgramState,
    /* private fields */
}

Fields
program_id: Uuidmy_idx: usizeparticipants: Vec<ParticipantData>leader: usizeprotocol: ProtocolTypestate: ProgramState
Implementations

impl Program

pub fn load( storage: Rc<Storage>, program_id: &Uuid, ) -> Result<Self, ProgramError>

pub fn im_leader(&self) -> bool

pub fn save(&self) -> Result<(), ProgramError>

pub fn setup( id: &Uuid, protocol_type: &str, peers: Vec<CommsAddress>, leader: usize, program_context: &mut ProgramContext, storage: Rc<Storage>, config: &ClientConfig, ) -> Result<Self, BitVMXError>

pub fn prepare_aggregated_keys( &mut self, context: &ProgramContext, ) -> Result<HashMap<String, PublicKey>, BitVMXError>

pub fn build_protocol( &mut self, context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn get_address_from_pubkey_hash( &self, pubkey_hash: &PubKeyHash, ) -> Result<CommsAddress, BitVMXError>

pub fn request_helper<T>( &mut self, program_context: &ProgramContext, to_send: Vec<(PubKeyHash, T)>, msg_type: CommsMessageType, ) -> Result<(), BitVMXError>
where T: Serialize + Clone,

pub fn send_keys( &mut self, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn receive_keys( &mut self, comms_address: CommsAddress, msg_type: CommsMessageType, data: Value, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn send_nonces( &mut self, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn receive_nonces( &mut self, comms_address: CommsAddress, msg_type: CommsMessageType, data: Value, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn send_signatures( &mut self, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn receive_signatures( &mut self, comms_address: CommsAddress, msg_type: CommsMessageType, data: Value, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn tick( &mut self, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn process_comms_message( &mut self, comms_address: CommsAddress, msg_type: CommsMessageType, data: Value, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn send_ack( &self, program_context: &ProgramContext, comms_address: CommsAddress, msg_type: CommsMessageType, ) -> Result<(), BitVMXError>

pub fn get_transaction_by_name( &self, program_context: &ProgramContext, name: &str, ) -> Result<Transaction, BitVMXError>

pub fn dispatch_transaction_name( &self, program_context: &ProgramContext, name: &str, ) -> Result<(), BitVMXError>

pub fn notify_news( &self, tx_id: Txid, vout: Option<u32>, tx_status: TransactionStatus, context: String, program_context: &ProgramContext, ) -> Result<(), BitVMXError>

pub fn get_transactions_to_monitor( &self, program_context: &ProgramContext, ) -> Result<(Vec<Txid>, Vec<(Txid, u32)>), BitVMXError>

pub fn move_program_to_next_state(&mut self) -> Result<(), BitVMXError>

This function should only be called when the program is in the correct state, otherwise it will transition to the next state at the wrong time and break the program’s flow

pub fn get_tx_by_id(&self, txid: Txid) -> Result<Transaction, BitVMXError>